outlet-service/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
exceptiongroup==1.2.2
fastapi==0.99.1
fastapi-jwt-auth==0.5.0
filelock==3.18.0
h11==0.14.0
identify==2.6.9
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==1.10.16
pydantic_core==2.33.0
PyJWT==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
